Backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
gunicorn==21.2.0
pytesseract==0.3.10
pillow==10.1.0
PyPDF2==3.0.1
langchain==0.1.0
chromadb==0.4.18
sentence-transformers==2.2.2
google-generativeai==0.3.1
python-dotenv==1.0.0
easyocr==1.7.0
numpy==1.24.3
pandas==2.0.3

